requirements.txt
anthropic==0.28.0
python-dotenv==1.0.0
gitpython==3.1.43
PyGithub==2.3.0
